Fix compiling tests, rename module looped_conter to wrapped_counter
